BACK/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
numpy==1.24.3
idna==3.4
mysqlclient==2.1.1
PyJWT==1.7.1
pytz==2023.3
requests==2.29.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
urllib3==1.26.15
